Added a port of the base58 Fast decoding and Trivial decoding
The port is based off of the github.com/trezor/trezor-crypto/base58.c code. There can be some more optimizations around buffering but it works with the tests.
Updated the tests to test Fast and Trivial decode as well
Updated the benchmarks to include decoding (seems to be just over a 6x improvement)